Refactor record sample module
